+++ datacontract/export/avro_converter.py
@@ -108,6 +108,8 @@ def to_avro_type(field: Field, field_name: str) -> str | dict:
     elif field.type in ["time"]:
         return "long"
     elif field.type in ["object", "record", "struct"]:
+        if field.config is not None and 'namespace' in field.config:
+            return to_avro_record(field_name ,field.fields ,field.description ,field.config['namespace'])
         return to_avro_record(field_name, field.fields, field.description, None)
     elif field.type in ["binary"]:
         return "bytes"